+++ packages/unexus-shell/include/usersettings.h
@@ -23,6 +23,7 @@ class UserSettings : public QObject
     Q_PROPERTY(QString updateChannel READ updateChannel WRITE setUpdateChannel NOTIFY updateChannelChanged)
     Q_PROPERTY(QString controlCenterSection READ controlCenterSection WRITE setControlCenterSection NOTIFY controlCenterSectionChanged)
     Q_PROPERTY(int notificationTimeoutSeconds READ notificationTimeoutSeconds WRITE setNotificationTimeoutSeconds NOTIFY notificationTimeoutSecondsChanged)
+    Q_PROPERTY(double notificationSilencedUntil READ notificationSilencedUntil WRITE setNotificationSilencedUntil NOTIFY notificationSilencedUntilChanged)
 
 public:
     explicit UserSettings(QObject *parent = nullptr);
@@ -43,6 +44,7 @@ public:
     QString updateChannel() const { return m_updateChannel; }
     QString controlCenterSection() const { return m_controlCenterSection; }
     int notificationTimeoutSeconds() const { return m_notificationTimeoutSeconds; }
+    double notificationSilencedUntil() const { return m_notificationSilencedUntil; }
 
 public slots:
     void setThemeIndex(int themeIndex);
@@ -61,6 +63,7 @@ public slots:
     void setUpdateChannel(const QString &channel);
     void setControlCenterSection(const QString &section);
     void setNotificationTimeoutSeconds(int seconds);
+    void setNotificationSilencedUntil(double silencedUntil);
 
 signals:
     void themeIndexChanged();
@@ -79,6 +82,7 @@ signals:
     void updateChannelChanged();
     void controlCenterSectionChanged();
     void notificationTimeoutSecondsChanged();
+    void notificationSilencedUntilChanged();
 
 private:
     QSettings m_settings;
@@ -98,4 +102,5 @@ private:
     QString m_updateChannel = "stable";
     QString m_controlCenterSection = "system";
     int m_notificationTimeoutSeconds = 7;
+    double m_notificationSilencedUntil = 0.0;
 };